backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
requests

python-jose[cryptography]
bcrypt==4.0.1
passlib==1.7.4

APScheduler

transformers
torch
sentence-transformers

geopy